Unit testing reorganization.
Unit testing was moved into its own directory as a place to make small
tests to functions without running the entire pipeline since that
takes a long time. It has a separate store of data specific to the unit tests
on Brad's s3 which can be updated/downloaded with the manage_unit_test_data.py
script. This added boto as a requirement.

Meant for tests of the 'does this function work correctly?' variety rather
than 'does this whole thing actually run?'