Skip "unknown" variants when doing somatic calls with VarScan
VarScan produces broken REF fields when processing "Unknown" records,
which are loci with complex rearrangements (e.g. gain in normal, loss in
tumor). Attempts to workaround the issues have just uncovered more
problems, so for now the solution is to skip these variants.

Once VarScan is fixed, the hacks will be dropped.

This handles issue #134.