Compare creation times of BAM and index files
As a result of errors or re-runs of the pipeline, the index files may
become out of date. If this is not properly checked, this will cause
further processing by Picard to fail with odd errors
(out of bound exceptions, etc.).

The logic is not the best, so suggestions are welcome.

This fixes issue #180.