Switch fastq detection to exclude impossible encodings.
This way it can abort early if it finds only one possible
encoding matches. Dropped max number of reads to look at to 1000,
so we aren't taking forever going through the files.
Added more tags to the unit tests so you can run all of the
RNA-seq tests at once.

Thanks to @abushoy for the nice implementation.