Update main.py
By default on Mac OS X the maximum number of files a process can open is 10,240. Unless the number 50,000 was chosen for a good reason, I suggest we change `target_procs` to `10240` to ensure a better out-of-the-box experience for Mac users.

The alternative would be to add a check for OS X before lowering `target_procs`.