Update parallel.rst
Added information on how to configure ``mem_free`` as a consumable resource, as by default the ``mem_free`` option only checks available memory upon spawning the engines. If we had a host with, say, 16 slots and 60 GB of memory and each job required 16GB, SGE would still happily spawn 16 engines on the host if ``mem_free`` wasn't a consumable.