Refactor: remove code referring to a lane
The concept of a lane to process is stale in the code, which was creating some
confusing code simply due to naming.  Sometimes it was lanes, lane_items,
samples, data, etc. This keeps 'samples' as a list of those data dictionaries
and removes the lane module.

Thanks to @kern3020 for bringing up the issue and @chapmanb for clarifying
that the code was stale.