Trimming overhaul, removal of decompression of FASTQ files.
1) Don't uncompress FASTQ files while trimming.
AlienTrimmer does not work on gzipped files, and the recent skewer paper shows
AlienTrimmer is not as sensitive as the other trimmers, so it is time to
retire it.

http://www.biomedcentral.com/1471-2105/15/182

2) cutadapt + skewer to trim
cutadapt has had the weakness of being slow and not handling paired data
particuarlly well. We run cutadapt without removing any reads and then use
scythe to remove the reads that are too short, streaming between the two so we
don't touch disk.  With the 1.4 release of cutadapt, trimming has been sped up
a ton, especially with gzipped files so this should be fast.