Fixes for matplotlib error.
When some functions are getting pickled and sent to the engines,
I think what is happening is that is the first time matplotlib
has been imported on the engine even though it has been imported
at the top-level bcbio-nextgen process. So we have to set it to
use 'Agg' as the backend. We also set it to force the creation
of a new backend, which might work to wipe out a stale matplotlib
backend that is looking for a display.

Also bonus trailing space removal issues.

Thanks to @chapmanb, @lpantano and @mistrm82 for figuring out what
was up. Unless this didn't fix it, then no thanks.