Use one consequence only with VEP if clinical_reporting is active
When looking for variants at the gene / exon level, sometimes it is
desirable to have only one consequence annotated, in particular for data
in clinical research. bcbio-nextgen already does this with snpEff
(-canonical option), but so far it was not used by VEP.

This patch adds this behavior (defaulting to false, hence not breaking
existing setups. It does so by using VEP's --pick option, introduced in
version 75.

From the VEP docs:

"Pick once line of consequence data per variant, including
transcript-specific columns. Consequences are chosen by the canonical,
biotype status and length of the transcript, along with the ranking of
the consequence type according to this table. This is the best method to
use if you are interested only in one consequence per variant. "