Add stringtie support for FPKM expression.
We're hoping to replace Cufflinks expression and assembly
with Stringtie, as it is way faster. This is the first step
of that process, we're hoping it can mostly be a drop in
replacement.

This is disabled for now while we figure out what is causing
some differences in the FPKM values calculated.