Improve transcriptome reconstruction performance.
1) fixes memoization of merged database
2) fix for getting the database where it was remaking the
   database if it already existed if in_memory was specified

Partially addresses #849.

We were making the reference database again in memory for no reason, so
we fixed that. This doesn't fix the underlying speed issue of making the
gffutils database; we could drop the merge features parameter and assume
all features are unique at this point since that should have been
handled by Cufflinks merge. The fix in daler/gffutils#48 might make this
moot though.

Thanks to @schelhorn for pointing this out.