Use disambiguated reads with aligning to the transcriptome.
This also removes the rsem: True flag, replacing it with
transcriptome_align: True. This has the meaning of turning
transcriptome-only aligning on, for use with RSEM and other
downstream tools requiring that. Not many people have
rsem: True turned on, so this deprecation won't affect
very many users so now is the time to do it.

Thanks to @schelhorn for doing all of the heavy lifting for
this patch.

Closes #859.