Provide version sort for un to date coreutils
Recent versions of GNU coreutils sort provide -V/--version-sort, which
does natural sorting of numbers in text so that it will correctly
sort chr1, chr2, instead of chr1, chr10, chr2. This checks for
support in the current sort and uses it if available. This also
works around a bug in coreutils 8.22 (fixes #1017)