Avoid race-conditions in BAM indexing
The BAM index check during callability processing can cause race
conditions on some shared filesystems where timestamps are out of date.
The check will think that the BAI file is old and try to re-index
in parallel, which can cause missing index errors as it gets re-created.
Since the index is be pre-prepared upstream this removes the
check here to avoid the problem.