slurm memory fixes for small AWS clusters
On smaller clusters, typically on AWS, bcbio will try to overschedule
memory due to the presence of the bcbio and iPython controller jobs
on the small number of nodes. Because those jobs are present, larger
engine jobs cannot get scheduled. This reduces memory targets for
these systems to work around the problem. Thanks to @guillermo-carrasco