Start on reboot of common workflow language support
Moves preparation logic for CWL output of the initial test CWL script
and into bcbio_nextgen, with an interface through bcbio_vm. The initial
preparation from bcbio YAML to CWL is not a good target for running
in CWL since the input file requires access to the filesystem and
relative paths. The outputs produced by `bcbio_vm.py cwl` will be CWL
files for actual runs.