Add Sailfish as the default non-count expression estimator.
This moves Cufflinks to an opt-in expression caller via the expression_caller entry
in the configuration and adds Sailfish as the default caller.

Sailfish expression values are added via four files. The first is the
Sailfish directory which can be used with Sleuth. The second is a
tidy combined.sf file that has all of the TPM calls for all transcripts
in all samples in tidy format. combined.isoform.sf.tpm has isoform-level
TPM in a matrix format and combined.gene.sf.tpm has gene-level
TPM.