Handle disambiguation with Sailfish.
This is an attempt to address #1132. If this turns
out to be poor we can do the more complicated
strategy of regenerating FASTQ files from the
disambiguated BAM files and re-quantitating. Thanks
to @schelhorn for raising the issue.

This commit also eliminate temporary directory issues that occur when
using tempfile. Thanks to @mjafin and @chapmanb for discussion about it.