Enable Stringtie transcriptome assembly.
As part of this work, this commit also allows multiple transcriptome
assemblers to be run on the same sample. The output from all
assemblers is merged together at the end with Cuffmerge.

This also deprecates the `assemble_transcriptome` option and
replaces it with a `transcript_assembler` option, which takes
a list of assemblers to run. (Supports `cufflinks` and `stringtie`
with this commit).