Add the client module
The client module tries to improve the readability of the project
by exposing a generic interface for the interaction with the
arguments received from the client and the actions that must be
taken.

This module will also be used in the bcbio-nextgen-vm in order to
avoid duplicated code across the projects.