rMATS
Hi Brad, Hi Roryl, Hi Lpantano,
I Brad I hope now is better.

@lpantano I have set rMATS to run using replicate for both sample and
control.

in the phenotype you have to set control or a sample name and I have
add replicate that as to be set as an integer.

I definitely agree with your and @roryk comment on False positive issues