Variant call samples with heterogeneous coverage
When we have different coverage profiles for batch of samples processed
together, bcbio would pick the coverage/variant_regions for the
first sample and use for processing. This leads to order dependence
and unexpected behavior when calling exome and WGS samples together
(#1326). To handle this case, we explicitly pick the sample with the
most genome coverage in a batch (or tumor in a paired somatic analysis)
and use this coverage/variant_regions for calling. In the case of
combining exome and WGS calls, exomes will have no calls in WGS only
regions.