Add more helpful error message when two samples are named exactly the same.
For example: data1/SRR566986.fastq.gz and data2/SRR566986.fastq.gz resolve to
the same stem, although they are under different directories. bcbio can't
handle this during the templating step, because it doesn't know which one to
assign to the samplename in the CSV file.

Thanks to @ssaif for pinging us about the error and providing a reproducible
example.