Add on-the-fly spikein quantification to RNA-seq pipeline.
Update documentation for the new `spikein_fasta` option.

Thanks to @lpantano for implementing and testing the spikein code.
She tested it against hg19 with and without aligning to the
genome/spikeins at the same time, and the counts match almost exactly,
so it is valid to do this separately.

Thanks to @saboswell for the spikein suggestion.

Closes #1575.