Docs: cutoff-based soft filtering instead of hard
There were confusing mentions in the code about hard filtering, which
referred to applying a cutoff-based soft filter (as opposed to methods
like VQSR). However, this was confusing in comparing hard (removing
variants) versus soft (flagging variants in the FILTER field) filters.
Fixed the language so it's reflective of what happens. #1684