Change cutadapt trimming to support the single pass option.
cutadapt 1.8+ can handle the paired files in one go and as a bonus does
some extra sanity checking of the FASTQ files as it goes. We swapped to
use that. A bonus is this speeds up trimming because it only runs
cutadapt once.

Also closes #1700 via https://github.com/bioconda/bioconda-recipes/pull/3252