PREP SAMPLES: allow full name comparison
when merging samples it could be the stem of files
are the same, but not the full path. This will give an issue
as it is because we expect all files to have different names.
If paired files are in the same folder, then we can compare
full names and it should work.

No compatible with:

* same file names in different folders
* paired files in different folders

In this case, or rename files, or move paired data to same folder.

Thanks @pengxiao78

https://github.com/chapmanb/bcbio-nextgen/issues/1802