Use `tools_on: [lumpy_usecnv]` for Lumpy/CNVkit
To feed CNVkit input as prior evidence to Lumpy, explicitly turn on
with `tools_on: [lumpy_usecnv]`. This changes the default, which
always ran CNVkit in a separate stage prior to Lumpy and caused longer
runtimes since they couldn't happen in parallel. Part of larger move
to use a separate explicit recalling step for SV calling instead
of this custom integration.