Upload: avoid double upload of population VCFs
More carefully manage output and naming of population VCFs with
introduction of vcfanno. population VCFs can be identical to sample VCFs
when we don't split samples and use only annotation and not
decomposition. In that case we only output population level VCF and
avoid samples. When we generate GEMINI databases and decompose
population VCFs we output both. Also fixes naming of output files
to be uniform. Thanks to @vladsaveliev