CWL: complex record parsing for all steps
Finalizes reworking on record parsing to use JSON inputs instead
of serializing to the command line. This works for QC and variant
calling as well as alignment, so the whole re-worked pipeline now
runs on cwltool and Toil.

Improves representation of batched records by collapsing the records
themselves, rather than items in the records. This enables more complex
batches and makes debugging outputs easier.