Include FreeBayes dbSNP and GATK annotated file
bcbio would generate annotated GATK and dbSNP files but not include
them in the merged output file, leaving the annotations behind. This
ensures the annotated files get passed. It also handles the backup
case where we want to annotate with dbSNP and GATK is not present.
Thanks to @semal for reporting. Fixes #1943