CWL: single file tarballs for complex collections
Avoids using complex secondaryFile representations for directories
of files like aligner indices, RTG and snpEff. Instead create a
tarball of files and pass to jobs as a single file. secondaryFiles
are handled differently across platforms and not supported on WDL, so
provides better portability. Also enables snpEff support for effects
prediction in CWL.