modify policy for prepending paths
With the previous policy prevented paths that were already in the
original list from being prepended to it, even if the existing paths
were not at the front of the list.  E.g., with the original policy
'/a' could not be prepended to '/b:/a', so it could never override the
leading '/b'.

The new policy fixes this problem.  From the docstring for _prepend:

    The list of paths represented in the returned value consists of
    the first of occurrences of each non-empty path in the list
    obtained by prepending the paths in to_prepend to the paths in
    original.