peddy: fix bug where not all files upload on first pipeline run
peddy was missing non-HTML report files during the first run of a pipeline
and generation of results, because the `peddyfiles` output was only calculated
before the run when files didn't exist. This fixes by calculating a single
time after the run so results are consistent on initial and re-runs. Thanks to
@vladsaveliev and @ohofmann