If R will be used downstream, sanitize the descriptions.
R can't handle names that start with numbers as row/column
names well. For pipelines that are likely to involve heavy R
usage downstream, we clean up the names here by putting an X
in front of the name if it starts with a number.

Thanks to Avinash Reddy and Dan Stetson at AstraZeneca and
Mike Steinbaugh and Brad Chapman for the discussion.