PureCN SV/heterogeneity: initial integration
Provides baseline support for running PureCN as an svcaller. Feeds CNVkit
segmentation and VCF file with germline variants (vardict) as inputs.

Work so far is only on tumor/normal case, where we'll plan to test on
real samples. Longer term need to work on tumor-only case and handling panels
of normals.

Discussion in #2469