#2520 fix logging in parallel threads (local multi-core)
In the transition from 0.11 to 0.12 joblib moved to a new default
backend ('loky') that does not work with the logging setup for
local parallel setups (multi-core) via logbook.Logger and handlers
communicating via a queue.

This fix enforces the previous default backend, 'multiprocessing',
which spawns several processes of the python interpreter and seems the
only parallel execution scheme that currently works w.r.t.
logging in a local parallel execution. See
https://media.readthedocs.org/pdf/joblib/latest/joblib.pdf and
https://github.com/joblib/joblib/blob/4f1ca521480713c403e1ddd362af43def6a059c8/joblib/parallel.py#L40.

Co-authored-by: gabeng <ben.liesfeld@limbus-medtec.com>