#2618: fix qc/variants/*_bcftools file creation
These files are not created anymore because get_active_vcinfo
returns an empty array instead of the first element on invocation
when running a non-ensemble setup.

Instead, it now fails 'gracefully' by attempting to find an entry for
the 'ensemble' variant caller, but if it does not find one it will still
return the first non-ensemble one.