Avoid use of apply for multicore function runs
Python 3 removes apply, so we need to adjust the bcbio.utils.map_wrap function
which uses is to double unpack into arguments (via both the `*` approach and
`apply`) which was necessary to enable arbitrary number of input arguments
to the original function. We do this by unpacking on the jobs calls that
use this in `run_multicore` and `runfn`.