Tumor-only prioritization: do not filter inputs with COSMIC support
When we had known database support of a variant (`EPR=pass,cosmic`)
this would result in filtering due to the second item not also being pass.
bcftools uses `or` instead of `and` logic to check multiple fields:

https://samtools.github.io/bcftools/bcftools.html#expressions

Since we put `pass` in the first column, we can check just that to avoid
the issue. Thanks to Moiz Bootwalla

https://groups.google.com/d/msg/biovalidation/Wfm9dPKRxkk/NMoP_cdHEwAJ