Ignore the installed tx2genefile.
This makes the logic when we provide custom GTF files easier.  The
tx2genefile is pretty low-cost to make, so making it each time from the
GTF file we are actually using makes more sense than trying to cache it.

This fixes an issue where we were not creating a tx2gene.csv file from
GTF files passed on with the `transcriptome_gtf` option.

Thanks to @lubinskibu and Avinash Reddy for the suggestion.