Remove missing ALT variants from GATK 4.1.0.0.
When a multiallelic variant is called, and one of the alleles has poor
quality GATK4 can output a variant with no ALT allele set.

We saw this in Freebayes as well, so we borrow that filter to remove
these variants as a workaround.

Thanks to @chapmanb, @ldgauthier and the GATK team for helping figure
out why this was happening, suggesting a workaround and ultimately
fixing the issue.

See broadinstitute/gatk#5650 for more information and the
ultimate resolution.