Allow pre-demultiplexed single-cell data.
Sometimes we get samples, especially from DGE experiments, where the
cells/wells have already been demultiplexed into individual FASTQ files.
This adds support for this case; the resulting columns are named via the
sample name in the metadata instead of after a cellular barcode.

Also fixes an issue where the installed GTF was not being used if one is
not specified.