MultiQc: add RNA-seq variantcalling metrics
This sets bcftools/snpEff as a multiqc function to be run, when RNA-seq
variantcalling is performed.

Currently, the structure of the RNA-seq variant calling datadict is
different than what the standard variant calling pipeline expects, so we
had to add some code to account for that.

It would be good to unify the two and document the format.

Closes #2641. Thanks to @WimSpee for the suggestion.