HLA: ensure calls with UMI processing inputs
Avoids skipping HLA calls with UMIs which have two sets of fastq
extractions: the original and consensus reads. We use the original
reads avoiding consensus reduction as part of OptiType since we'd
ideally like to feed as much information as possible and it's not
clear consensus calling will help in highly polymorphic regions.