Flip strandedness of qualimap.
I think this is either a bug in qualimap or my misunderstanding of
what the qualimap option means.

The misunderstanding could be coming from the read being from the RNA
but the RNA is already on the opposite strand of the feature, so when we
have reverse-forward reads, the reverse read is actually in the right
direction of the feature, just not the RNA.

Either that or qualimap has the options flipped. In either case, this
fix fixes the problem.