[PoC] Use Mutect2's new read model for filtering
According to the GATK's doc, Mutect2 can detect read orientation
artifacts (such as stand bias caused by FFPE, or by Illumina Novaseq)
and improve results.

This is a proof of concept (completely untested at this point). In
particular this is posted for early review.

Some important points to act on:

- Do we want this on at all times (GATK says it won't hurt performance)?
- How do we make sure this is only run for paired-end samples (see
FIXME)?
- Is the current simplistic logic for file names acceptable?

This will fix issue #2846.