Force unstranded for qualimap with RNA-seq data
We use qualimap mostly to look at the overlap of genomic features for
RNA-seq, and qualimap returns incorrect results when looking at hisat2
alignments when run in stranded mode. With the same firststrand
protocol, setting the stranded mode to forward worked with one dataset
and another dataset needed it to be set to reverse.  I think hisat2
might be mucking up the flag somehow, but I haven't been able to figure
out why.

Just running in unstranded mode gives results close to the stranded
mode, so we will do that for now.