Support for Salmon with STAR alignments.
Rob's group has a preprint out showing that for some cases, it is better
to quantitate on the whole genome alignments instead of just the
transcriptome. This adds support for that, with the
`quantify_genome_alignments` option.

This does the following:

1. Add --validateMappings to Salmon read quantification
2. Add a `quantify_genome_alignments` option; if set to True, this
quantiates from the whole genome alignments that have been
translated to the genomic coordinates, not the reads.

This works ONLY with STAR as the aligner, currently, so we can't use
this with hg38. We will be adding an option to do the SA mode in Salmon
which will take decoys in order to be able to handle hg38.  We're also
implementing hg38 support for STAR in bcbio by stripping out the ALTS
alleles before making the index.

This is functional right now and available for folks to try out.