Explicitly set locale to support UTF-8.
Addresses #2890. Useful discussion is in that issue as well.  This might
need to get reverted if it breaks a lot of stuff for folks.

We prefer C.UTF-8 if it is available. If not then we fall back to
whatever the first locale on `locale -a` is set. C.UTF-8 is the default
locale if `locale` is not availble for some reason.