Support of correct_umis for a pre-defined set of barcodes (#2949)
This adds support for adding a whitelisted set of UMIs similar to the way we provide for a cellular barcode whitelist. The flag to do this is `correct_umis`, which takes a file of whitelisted UMIs. This is useful for some protocols that use a restricted set of UMIs.

This also includes a few documentation fixes that came up while testing:

1) Clarified how to create a testing environment
2) Clarified how `min_allele_fraction` works, it is confusing that fraction here is a percent.