Manual STAR 2-pass basic.
STAR 2-pass basic takes the per-sample junction file and re-runs STAR
with it. We can't use 2-pass basic because it doesn't work properly when
using the input piping we are doing to get around filesystem issues
when using the `--readFilesCommand` flag. So here we just manually
do it, running STAR twice using the junction file.