Add `buildversion` option.
This lets `bcbio_setup_genome.py` pass on an arbitrary string to
save as the gene build version when setting up a genome. The text
is freeform, but the suggested format is source_version. Examples
are Ensembl_75, EnsemblMetazoa_45, Flybase_25, etc. This lets us
track better where gene builds came from.