Ensure EBV coverage is calculated.
For hg38, we include EBV as a chromosome, so we skip quantitating it
when looking at viral coverage. Here we add the missing EBV back in if
it is included as an artifical chromosome of the host genome.

GRCh37 and hg19 was also missing EBV, we fixed this by updating
the indices from GDC to include EBV here:

https://github.com/chapmanb/cloudbiolinux/commit/1a00669705b48b7e8f565890fb40f3086544dcc2