With > 100 samples, skip broken bcbioRNASeq plots
bcbioRNASeq has a few plots that color the points based on either the
samplename or interesting_groups. If interesting_groups is not set, and
there are many samples, the coloring fails. This checks if there are
over 100 samples and if interesting groups is not set, and skips running
the broken plots if they are not set properly.