Support single-end reads for ATAC-seq.
While not ideal, there are some datasets for ATAC-seq that are single
end. These fixes make it so we can handle these datatypes.

The main problem with single-end data is we can no longer estimate the
NF/MN/DN/TN fraction and have to call peaks on the full BAM files. We
are also unable to calculate the ENCODE quality control metrics, nor
does ataqv produce as many useful metrics.