Add `--enumerate` flag to OptiType.
This will return the top 10 matches for the HLA type from the data,
which can be helpful when deciding how confident we should be in the
calls.

I ran this on the NA12878 trio and it called the same HLA types as
this paper
(https://bmcgenomics.biomedcentral.com/articles/10.1186/1471-2164-16-S2-S7/tables/2),
but the scores were pretty minor between the top and second highest
hits.

See https://github.com/FRED-2/OptiType/issues/89 for a discussion of
why we turned on the `--enumerate` option.