Workaround for htsjdk not supporting VCF4.3.
VarDict emits variant files following the VCF4.3 spec, but htsjdk cannot
write VCF4.3 so it throws an error when it encounters a VCF file
reporting this version.

Most of the 4.3 changes seem to revolve around optional tags, so here we
just edit the reported version to be 4.2 which should work fine for now.

See https://samtools.github.io/hts-specs/VCFv4.3.pdf section 7.2 to
compare the VCFv4.2 and VCFv4.3 changes.

Closes #3339. Thanks to @kdrichards for pointing out the issue.