SCRAMble integration (#3341)
This adds support for the mobile element caller SCRAMble (https://github.com/GeneDx/scramble). This is designed to be used on exome sequencing but we have tried it on whole genome sequencing and it does okay.

The license means this cannot be used for commercial enterprise without permission from them (https://github.com/GeneDx/scramble/blob/master/LICENSE). The alternative MELT has a much more draconian license so we went with SCRAMble.

Thanks to @chanzuckerberg for funding to support this feature and to @hackdna for implementing and validating it.