Enable consensus calling for broad peaks.
Thanks to @mistrm82 and @yoonsquared for pointing out was is not
enabled.

This changes the consensus peak calling for narrowPeaks for ATAC-seq and
ChIP-seq. Before we were finding summits and expanding peaks out to 250
bases for each peak, now we don't do that and just keep the size of the
peak with the highest score.